+++ src/main/java/com/example/api/domain/entity/Address.java
@@ -1,2 +1,50 @@
-package com.example.api.domain.entity;public class Address {
+package com.example.api.domain.entity;
+
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.validation.constraints.NotEmpty;
+
+@Getter
+@Setter
+@Entity
+@Table(name = "ADDRESS")
+public class Address {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    @NotEmpty
+    private String street;
+
+    @Column(nullable = false)
+    @NotEmpty
+    private String city;
+
+    @Column(nullable = false)
+    @NotEmpty
+    private String state;
+
+    @Column(nullable = false)
+    @NotEmpty
+    private String zipCode;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "customer_id", nullable = false)
+    @JsonIgnore
+    private Customer customer;
+
 }
